minor changes in tests

+++ tests/TimeHelper-tests.php
@@ -9,7 +9,7 @@ $format = TimeHelper::TIME_IN_HOURS_SHORT;
 echo TimeHelper::secondsToTime(12345, $format) . PHP_EOL;
 
 $format = TimeHelper::TIME_IN_HOURS;
-echo TimeHelper::secondsToTime(3600 + 185, $format) . PHP_EOL;
+echo TimeHelper::secondsToTime((int)(3600 + 185), $format) . PHP_EOL;
 
 TimeHelper::$hours = "heures";
 TimeHelper::$minutes = "minutes";
